described parameters | applied codeguides
