+++ panther_docking/config/panther_docking_server.yaml
@@ -29,14 +29,14 @@
       type: panther_charging_dock
       frame: <robot_namespace>/map
       dock_frame: main_wibotic_receiver_requested_pose_link
-      pose: [1.0, 1.5, 1.57] # position of the dock device (not the staging position), the front (X+) of the dock should point away from the robot
+      pose: [1.0, 1.5, 1.57] # position of the dock on the map. Used also for spawning dock in the simulation.
       apriltag_id: 0
 
     backup:
       type: panther_charging_dock
       frame: <robot_namespace>/map
       dock_frame: backup_wibotic_receiver_requested_pose_link
-      pose: [-1.0, 1.5, 1.57] # position of the dock device (not the staging position), the front (X+) of the dock should point away from the robot
+      pose: [-1.0, 1.5, 1.57]  # position of the dock on the map. Used also for spawning dock in the simulation.
       apriltag_id: 1
 
     controller:
